+++ src/lib_client_base_unix/client_config.ml
@@ -1342,3 +1342,5 @@ let clic_commands ~base_dir:_ ~config_commands ~builtin_commands ~other_commands
   config_commands @ builtin_commands @ other_commands
 
 let logger = None
+
+let advertise_log_levels = None